fix: bound what a download accepts from the server
Four separate gaps, all on server-supplied input.

The sticker body was read into memory with no ceiling, so a response
that never ends took the process with it and surfaced as a bare Sticker
save failed. There is a 24 MB cap now.

RemoteMedia cast every URL to HttpURLConnection. A non-HTTP one threw
ClassCastException, which is not an IOException, so it escaped the mirror
loop and the remaining mirrors were never tried; it also skipped the
disconnect. Those entries are skipped and recorded instead.

The completeness check used getContentLength, an int that answers -1 past
2 GB, which is the size worth checking. It reads the long now.

Filenames were bounded at 160 characters. 160 Japanese characters is 480
bytes of UTF-8, past the 255 ext4 and f2fs take, and cutting at a fixed
index can split a surrogate pair. The cut lands on a code point and the
byte length is checked too.

Also widened the animated WebP frame bounds check to long, where two
large offsets could wrap past it.

+++ extensions/tiktok/src/main/java/app/morphe/extension/tiktok/download/AnimatedWebpGifConverter.java
@@ -162,7 +162,8 @@ final class AnimatedWebpGifConverter {
             int yOffset
     ) {
         if (frameWidth <= 0 || frameHeight <= 0 || xOffset < 0 || yOffset < 0
-                || xOffset + frameWidth > canvasWidth || yOffset + frameHeight > canvasHeight) {
+                || (long) xOffset + frameWidth > canvasWidth
+                || (long) yOffset + frameHeight > canvasHeight) {
             throw new IllegalStateException("Animated WebP frame is outside its canvas");
         }
     }
